+++ deepcell/layers/__init__.py
@@ -43,6 +43,7 @@ from deepcell.layers.normalization import ImageNormalization2D
 from deepcell.layers.normalization import ImageNormalization3D
 from deepcell.layers.pooling import DilatedMaxPool2D
 from deepcell.layers.pooling import DilatedMaxPool3D
+from deepcell.layers.temporal import Comparison, DeltaReshape, Unmerge
 from deepcell.layers.temporal import TemporalMerge, TemporalUnmerge
 from deepcell.layers.tensor_product import TensorProduct
 from deepcell.layers.padding import ReflectionPadding2D
