+++ deepcell/layers/temporal.py
@@ -32,6 +32,58 @@ import tensorflow as tf
 from tensorflow.keras.layers import Layer
 
 
+class Comparison(Layer):
+    """Layer for comparing two sequences of inputs."""
+    def call(self, inputs):
+        x = inputs[0]
+        y = inputs[1]
+
+        x = tf.expand_dims(x, 3)
+        multiples = [1, 1, 1, tf.shape(y)[2], 1]
+        x = tf.tile(x, multiples)
+
+        y = tf.expand_dims(y, 2)
+        multiples = [1, 1, tf.shape(x)[2], 1, 1]
+        y = tf.tile(y, multiples)
+
+        return tf.concat([x, y], axis=-1)
+
+
+class DeltaReshape(Layer):
+    """Reshape changes between current and future frames"""
+    def call(self, inputs):
+        current = inputs[0]
+        future = inputs[1]
+        current = tf.expand_dims(current, axis=3)
+        multiples = [1, 1, 1, tf.shape(future)[2], 1]
+        output = tf.tile(current, multiples)
+        return output
+
+
+class Unmerge(Layer):
+    """Unmerge temporal inputs"""
+    def __init__(self, track_length, max_cells, embedding_dim, **kwargs):
+        super(Unmerge, self).__init__(**kwargs)
+        self.track_length = track_length
+        self.max_cells = max_cells
+        self.embedding_dim = embedding_dim
+
+    def call(self, inputs):
+        new_shape = [-1, self.track_length, self.max_cells, self.embedding_dim]
+        output = tf.reshape(inputs, new_shape)
+
+        return output
+
+    def get_config(self):
+        config = {
+            'track_length': self.track_length,
+            'max_cells': self.max_cells,
+            'embedding_dim': self.embedding_dim
+        }
+        base_config = super(Unmerge, self).get_config()
+        return dict(list(base_config.items()) + list(config.items()))
+
+
 class TemporalMerge(Layer):
     """Layer for merging the time dimension of a Tensor.
 
@@ -47,6 +99,13 @@ class TemporalMerge(Layer):
         b = inputs[1]
         new_shape = [-1, tf.shape(b)[2], self.encoder_dim]
         return tf.reshape(a, new_shape)
+
+    def get_config(self):
+        config = {
+            'encoder_dim': self.encoder_dim,
+        }
+        base_config = super(TemporalMerge, self).get_config()
+        return dict(list(base_config.items()) + list(config.items()))
 
 
 class TemporalUnmerge(Layer):
@@ -64,3 +123,10 @@ class TemporalUnmerge(Layer):
         b = inputs[1]
         new_shape = [-1, tf.shape(b)[1], tf.shape(b)[2], self.encoder_dim]
         return tf.reshape(a, new_shape)
+
+    def get_config(self):
+        config = {
+            'encoder_dim': self.encoder_dim,
+        }
+        base_config = super(TemporalUnmerge, self).get_config()
+        return dict(list(base_config.items()) + list(config.items()))
