+++ deepcell/model_zoo/tracking.py
@@ -45,6 +45,7 @@ from tensorflow.keras.layers import BatchNormalization, Lambda
 from tensorflow.keras.regularizers import l2
 
 from deepcell.layers import ImageNormalization2D
+from deepcell.layers import Comparison, DeltaReshape, Unmerge
 from deepcell.layers import TemporalMerge, TemporalUnmerge
 
 from spektral.layers import GCSConv
@@ -205,55 +206,6 @@ def siamese_model(input_shape=None,
     model = Model(inputs=inputs, outputs=final_layer)
 
     return model
-
-
-class Comparison(tf.keras.layers.Layer):
-    def call(self, inputs):
-        x = inputs[0]
-        y = inputs[1]
-
-        x = tf.expand_dims(x, 3)
-        multiples = [1, 1, 1, tf.shape(y)[2], 1]
-        x = tf.tile(x, multiples)
-
-        y = tf.expand_dims(y, 2)
-        multiples = [1, 1, tf.shape(x)[2], 1, 1]
-        y = tf.tile(y, multiples)
-
-        return tf.concat([x, y], axis=-1)
-
-
-class DeltaReshape(tf.keras.layers.Layer):
-    def call(self, inputs):
-        current = inputs[0]
-        future = inputs[1]
-        current = tf.expand_dims(current, axis=3)
-        multiples = [1, 1, 1, tf.shape(future)[2], 1]
-        output = tf.tile(current, multiples)
-        return output
-
-
-class Unmerge(tf.keras.layers.Layer):
-    def __init__(self, track_length, max_cells, embedding_dim, **kwargs):
-        super(Unmerge, self).__init__(**kwargs)
-        self.track_length = track_length
-        self.max_cells = max_cells
-        self.embedding_dim = embedding_dim
-
-    def call(self, inputs):
-        new_shape = [-1, self.track_length, self.max_cells, self.embedding_dim]
-        output = tf.reshape(inputs, new_shape)
-
-        return output
-
-    def get_config(self):
-        config = {
-            'track_length': self.track_length,
-            'max_cells': self.max_cells,
-            'embedding_dim': self.embedding_dim
-        }
-        base_config = super(Unmerge, self).get_config()
-        return dict(list(base_config.items()) + list(config.items()))
 
 
 class GNNTrackingModel(object):
